pyproject.toml
[project]
name = "vldmcp"
description = "A distributed (FoaF) MCP server using veilid and podman"
version = "0.0.1"
authors = [
    { name = "Gareth Davidson", email = "gaz@bitplane.net" }
]
readme = "README.md"
requires-python = ">=3.11"


dependencies = [
    "base32hex",
    "base58",
    "blake3",
    "click",
    "fastapi",
    "fastmcp",
    "gitpython",
    "mnemonic",
    "pydantic",
    "pydantic-settings",
    "pynacl",
    "psutil",
    "sqlmodel",
    "tomli-w",
    "veilid"
]

[project.optional-dependencies]
# not pydoc-markdown, circular dep
dev = [
    "pre-commit",
    "pytest",
    "coverage",
    "pytest-cov",
    "build",
    "twine",
    "ruff"
]

[project.urls]
Homepage = "https://bitplane.net/dev/python/vldmcp"
Repository = "https://github.com/bitplane/vldmcp"

[project.scripts]
vldmcp = "vldmcp.cli:cli"
vldmcpd = "vldmcp.daemon:main"

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true
